Move KeysProvider to a separate file
